added more checks
Needed to add some checks here. sqlite3 check would actually create an empty file if it didn't already exist and report it as "ok". To work around this we check if the file exists first AND then proceed to check it for errors. Also, added an email check for the root user so it can send an email, if not it will output the db errors to just the shell.